JSON recipes finished. Added images and folder. Firebase v1
