+++ .env
@@ -1,0 +1,13 @@
+# Configuración de firebase y firestore pitayapp-eoi
+API_KEY=AIzaSyCoQsUSujAxn2wZbrj812d8LRMuPvqKD9A
+PROJECT_ID=test-pitayapp
+APP_ID=1:899191902557:web:9cdf268e2ee25c6f4358b4
+AUTH_DOMAIN=test-pitayapp.firebaseapp.com
+FB_URL=
+SENDER=
+
+# Supabase
+SUP_URL=
+SUP_RTB=
+SUP_API=
+SUP_SVC=
